feat: Add GraphQL schema and resolver for calculating degrees of separation between actors
This change introduces a new GraphQL schema and resolver to calculate the degrees of separation between two actors. It utilizes the Neo4j graph database and the shortestPath function to find the shortest path of relationships between the two actors. The result is then processed and returned as a Path object containing the nodes traversed in the path. This feature will allow users to easily determine the connection between two actors in the context of the movie database.

References: #123